backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic[email]==2.10.3
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pywebpush==2.0.1
pytesseract==0.3.13
Pillow==11.1.0
python-dateutil==2.9.0
APScheduler==3.11.0
openai>=1.40.0
python-multipart>=0.0.9
